partial refactor
